@@ -1,5 +1,5 @@
 module github.com/things-go/cert4go
 
-go 1.15
+go 1.16
 
-require github.com/stretchr/testify v1.7.0
+require golang.org/x/crypto v0.0.0-20210921155107-089bfa567519
